requirements.txt
PyYAML==6.0.1
jsonschema==4.22.0
deepdiff==7.0.1
click==8.1.7